HGHMM-main/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic>=2.5.2
cryptography==41.0.5
mysqlclient==2.2.0
netifaces==0.11.0
distro==1.8.0